Extract InteractionMenu into its own component
